recommender/requirements.txt
requests
beautifulsoup4
sentence-transformers
qdrant-client
python-dotenv
